Resolve conflicts across five files:
- Docs/UltraCanvas/CHANGELOG.md — prepend UltraNet 0.3.0 entry above
  main's 0.2.32 head; keep every 0.2.x fix/merge intact.
- CMakeLists.txt — bump project VERSION to 0.3.0; concatenate the
  demo-source lists (UltraNetworkingExamples plus the new
  Slideshow/Album/MediaViewer/…/Video entries); take main's
  self-contained Tests wiring (drops Catch2 dependency).
- UltraCanvas/CMakeLists.txt — bump ULTRACANVAS_VERSION to 0.3.0;
  keep the UltraNet options block *and* main's GL comment block.
- UltraCanvas/core/UltraCanvasApplication.cpp — keep PostToUIThread /
  GetCurrent / ProcessPostedTasks; drop the old kDejaVuAllFonts array in
  favour of main's kEmbeddedAllFonts (Ubuntu + OpenSans).
- .github/workflows/build.yml — take main's cleaner trigger comment;
  merge every apt/brew/pacman list (UltraNet needs libcurl + c-ares,
  main added zbar + mupdf + gstreamer + mupdf third-party); fold
  ULTRACANVAS_BUILD_NET_TESTS=ON into both Configure steps so the tee
  configure.log flow (for the Verify OpenGL backend step) still works.

Verified: UltraNet suite 102 passed, 0 failed, 4 skipped (WS loopback
needs libcurl --enable-websockets) — same as pre-merge.

+++ Apps/DemoApp/UltraCanvasAutoCompleteExamples.cpp
@@ -16,14 +16,14 @@
 namespace UltraCanvas {
 
 // Helper function to create section separator
-    std::shared_ptr<UltraCanvasContainer> CreateAutoCompleteSeparator(long x, long y, long width) {
+    std::shared_ptr<UltraCanvasContainer> CreateAutoCompleteSeparator(float x, float y, float width) {
         auto separator = std::make_shared<UltraCanvasContainer>("ACSep" + std::to_string(x), x, y, width, 2);
         separator->SetBackgroundColor(Color(200, 200, 200, 255));
         return separator;
     }
 
 // Helper function to create section title
-    std::shared_ptr<UltraCanvasLabel> CreateAutoCompleteSectionTitle(long x, long y, const std::string& text) {
+    std::shared_ptr<UltraCanvasLabel> CreateAutoCompleteSectionTitle(float x, float y, const std::string& text) {
         auto title = std::make_shared<UltraCanvasLabel>("ACSecTitle" + std::to_string(x), x, y, 400, 25);
         title->SetText(text);
         title->SetFontSize(14);
